reorganize pivmandata
